RequireRequestDelete

+++ Kartverket.MetadataEditor/Controllers/MetadataController.cs
@@ -19,6 +19,8 @@ using Kartverket.MetadataEditor.Helpers;
 using Kartverket.MetadataEditor.Models.Translations;
 using Kartverket.MetadataEditor.Models.InspireCodelist;
 using GeoNetworkUtil = Kartverket.MetadataEditor.Util.GeoNetworkUtil;
+using System.Net.Mail;
+using System.Text;
 
 namespace Kartverket.MetadataEditor.Controllers
 {
@@ -1039,6 +1041,9 @@ namespace Kartverket.MetadataEditor.Controllers
 
             if (HasAccessToMetadata(model))
             {
+                if(RequireRequestDelete(model))
+                    return View("RequestDelete", model);
+                else
                 return View(model);
             } else {
                 return new HttpUnauthorizedResult();
@@ -1051,6 +1056,9 @@ namespace Kartverket.MetadataEditor.Controllers
         {
             MetadataViewModel model = _metadataService.GetMetadataModel(uuid);
 
+            if(RequireRequestDelete(model))
+                return new HttpUnauthorizedResult();
+
             if (HasAccessToMetadata(model))
             {
                 _metadataService.DeleteMetadata(model, GetUsername(), comment);
@@ -1062,6 +1070,70 @@ namespace Kartverket.MetadataEditor.Controllers
             {
                 return new HttpUnauthorizedResult();
             }
+        }
+
+        [Authorize]
+        [HttpPost]
+        public ActionResult RequestDelete(string uuid, string title, string comment)
+        {
+            SendEmail(uuid, title, comment);
+            TempData["Message"] = "Sendt anmodning om å slette " + title;
+            return RedirectToAction("Index");
+        }
+
+        private void SendEmail(string uuid, string title, string comment)
+        {
+            var fromEmail = ClaimsPrincipal.Current.GetUserEmail();
+
+            var nameFull = ClaimsPrincipal.Current.GetUserFullName();
+
+            if (string.IsNullOrEmpty(fromEmail))
+                fromEmail = System.Web.Configuration.WebConfigurationManager.AppSettings["WebmasterEmail"];
+
+            var message = new MailMessage();
+            message.To.Add(new MailAddress(System.Web.Configuration.WebConfigurationManager.AppSettings["WebmasterEmail"]));
+            message.From = new MailAddress(fromEmail);
+            message.Subject = "Anmodning om sletting av metadata  " + title;
+            StringBuilder b = new StringBuilder();
+            b.Append("Hei!<br/>\r\n");
+            b.Append("<p>Ønsker å slette metadata "+title+" med uuid: "+uuid+"</p>");
+            b.Append("Kommentar:<br/>\r\n");
+            b.Append(comment);
+            if(!string.IsNullOrEmpty(nameFull))
+                b.Append("<br/><br/>Mvh.<br/>" + nameFull);
+
+            message.Body = b.ToString();
+            message.IsBodyHtml = true;
+
+
+            using (var smtp = new SmtpClient())
+            {
+                smtp.Host = System.Web.Configuration.WebConfigurationManager.AppSettings["SmtpHost"];
+
+                try
+                {
+                    smtp.Send(message);
+                }
+                catch (Exception excep)
+                {
+                    Log.Error(excep.Message);
+                    Log.Error(excep.InnerException);
+                    throw new Exception("Sending av epost feilet");
+                }
+                Log.Info("Send email to:" + message.To.ToString());
+                Log.Info("Subject:" + message.Subject);
+                Log.Info("Body:" + message.Body);
+            }
+
+        }
+
+        private bool RequireRequestDelete(MetadataViewModel model)
+        {
+            var created = model.DateCreated;
+            if(!UserHasMetadataAdminRole() && (created.HasValue && created.Value != DateTime.Today))
+                return true;
+
+            return false;
         }
 
         private bool HasAccessToMetadata(MetadataViewModel model)
